Jelly: Show a Snackbar when removing an HistoryItem
This allows users to re-add the entry if deleted accidentally.

Also remove the long-press-to-delete functionality for history items, as
it would be quite a bit of additional effort to implement the SnackBar
for it and multiple ways of doing the same thing seem redundant anyway.

Change-Id: I14bea962e67bf0e3a9af67ef4cd141cbf322e3dd
(cherry picked from commit 1efae6f80c1382ae135e33f2392d3cdc06b0a338)